refactor: add separate file for types
