+++ src/components/view/sources/sources.ts
@@ -1,10 +1,6 @@
+import { Source } from '../../../types/types';
 import { View } from '../news/news';
 import './sources.css';
-
-export interface Source {
-  id: string;
-  name: string;
-}
 
 class Sources implements View<Source> {
   draw(data: Source[]): void {
