+++ src/types/types.ts
@@ -1,0 +1,44 @@
+export type HtmlEl = HTMLElement | null;
+export type Options = {
+  apiKey?: string;
+  sources?: string;
+};
+export type DrawFunc = (data: ApiResponse) => void;
+export type cb = (data?: ApiResponse) => void;
+export type GetRespOptions = {
+  endpoint: string;
+  options?: Options;
+};
+export type Response = { status: number; ok: boolean; statusText: string };
+
+export interface ILoader {
+  getResp: (obj: GetRespOptions, callback: cb) => void;
+  errorHandler: (res: Response) => Response;
+  makeUrl: (options: Options, endpoint: string) => string;
+  load: (method: string, endpoint: string, callback: cb, options: Options) => void;
+}
+
+export interface Controller {
+  getSources: (cb: cb) => void;
+  getNews: (e: Event, cb: cb) => void;
+}
+
+export interface NewsItem {
+  title: string;
+  description: string;
+  source: Source;
+  publishedAt: string;
+  author: string;
+  url: string;
+  urlToImage: string;
+}
+
+export interface Source {
+  id: string;
+  name: string;
+}
+
+export interface ApiResponse {
+  articles?: NewsItem[];
+  sources?: Source[];
+}
